add upstash ratelimiter

+++ src/server/queries.ts
@@ -6,6 +6,7 @@ import { redirect } from "next/navigation";
 import analyticsServerClient from "./analytics";
 import posthog from "posthog-js";
 import { revalidatePath } from "next/cache";
+import { ratelimit } from "./ratelimit";
 
 export async function getImages() {
 	const images = await db.query.images.findMany({
@@ -26,6 +27,9 @@ export async function getImage(id: number) {
 }
 
 export async function deleteImage(id: number) {
+	const { success } = await ratelimit.limit(posthog.get_distinct_id());
+	if (!success) throw new Error("Rate limited!");
+
 	await db.delete(images).where(eq(images.id, id));
 
 	analyticsServerClient.capture({
